requirements.txt
six
Django
channels
cssutils
gunicorn
psycopg2
requests
html5lib
multidict
django-csp
whitenoise
calmjs.parse
python-dotenv
beautifulsoup4
dj-database-url
django-recaptcha
uvicorn[standard]
django-encrypted-model-fields
git+https://github.com/lengors/cutils.git#egg=cutils
